requirements.txt
PyQt5
spotipy
requests
Pillow
